+++ Example/MultiSliderDemo/ViewController.swift
@@ -40,6 +40,9 @@ class ViewController: UIViewController {
         horizontalMultiSlider.showsThumbImageShadow = false
         view.addConstrainedSubview(horizontalMultiSlider, constrain: .leftMargin, .rightMargin, .bottomMargin)
         view.layoutMargins = UIEdgeInsets(top: 32, left: 32, bottom: 32, right: 32)
+
+        multiSlider.keepsDistanceBetweenThumbs = false
+        horizontalMultiSlider.keepsDistanceBetweenThumbs = false
     }
 
     @objc func sliderChanged(_ slider: MultiSlider) {