+++ Package.swift
@@ -5,17 +5,17 @@ import PackageDescription
 let package = Package(
     name: "MultiSlider",
     platforms: [
-        .iOS(.v9)
+        .iOS(.v9),
     ],
     products: [
-        .library( name: "MultiSlider", targets: ["MultiSlider"])
+        .library(name: "MultiSlider", targets: ["MultiSlider"]),
     ],
     dependencies: [
         .package(url: "https://github.com/yonat/SweeterSwift", from: "1.0.2"),
-        .package(url: "https://github.com/yonat/AvailableHapticFeedback", from: "1.0.2")
+        .package(url: "https://github.com/yonat/AvailableHapticFeedback", from: "1.0.2"),
     ],
     targets: [
-        .target(name: "MultiSlider", dependencies: ["SweeterSwift", "AvailableHapticFeedback"], path: "Sources")
+        .target(name: "MultiSlider", dependencies: ["SweeterSwift", "AvailableHapticFeedback"], path: "Sources"),
     ],
     swiftLanguageVersions: [.v5]
 )