+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+- when two thumbs overlap (`keepsDistanceBetweenThumbs == false`) let user drag lower thumb and not only upper thumb.
+
 ## [1.9.7] - 2019-08-22
 
 ### Added